Nieuwe bugs, bezig met leaderboard
