+++ src/StyleSheets/App.css
@@ -1,7 +1,7 @@
 /*Loads the 2 fonts the site from the Archeon uses contained in the Fonts directory*/
 @font-face {
     font-family: "Triplex-Light";
-    src: url("./Local_Files/Fonts/Triplex\ Light.ttf");
+    src: url("../Local_Files/Fonts/Triplex\ Light.ttf");
 }
 
 button {
@@ -29,4 +29,7 @@ button:hover {
 }
 .CorrectAnswer {
     color: #457c1f;
+}
+.ExtraInfo {
+    color :#fff;
 }
